Assign a z-index to the last-updated/hacker-clock.
Sometimes the last digit of the time (<p id="hacker-clock">) is behind the last updated field
(<p id="last-updated"), sometimes not. This is because the z-index css was not set for neither
of the two p-Elements. Thus the browser could either show hacker-clock before last-updated or
versus visa. (So sometimes, but not always, the browser showed a part of the last digit in the
hacker-clock behind the last-updated <p>.)

In order to get a uniform, good looking behaviour show the hacker-clock before last-updated.
Thus it is guaranteed that the last digit of the hacker-clock p-Element is always completely
visible.